+++ backend/app/crud/service.py
@@ -1,7 +1,7 @@
 from fastapi import HTTPException
 
 from db import get_db_client
-from models import Service, Product
+from models import Service, ServiceCreate, ServiceInput, ServiceInputCreate, Product
 from core import SETTINGS
 from crud import ProductCrud
 
@@ -33,7 +33,7 @@ class ServiceCrud:
         return Service.model_validate(response.data[0])
     
     @classmethod
-    async def create_service(cls, service_data: Service) -> Service:
+    async def create_service(cls, service_data: ServiceCreate) -> Service:
         """Create a new service."""
         
         client = await get_db_client()
@@ -50,6 +50,10 @@ class ServiceCrud:
         """Update an existing service."""
         
         client = await get_db_client()
+        
+        # check if service exists
+        if not await cls.exist_service_by_id(service_id):
+            raise HTTPException(detail="Service not found", status_code=404)
 
         response = await client.table(SETTINGS.service_table).update(fields).eq("id", service_id).execute()
 
@@ -63,6 +67,10 @@ class ServiceCrud:
         """Delete a service by ID."""
         
         client = await get_db_client()
+        
+        # check if service exists
+        if not await cls.exist_service_by_id(service_id):
+            raise HTTPException(detail="Service not found", status_code=404)
 
         response = await client.table(SETTINGS.service_table).delete().eq("id", service_id).execute()
 
@@ -80,7 +88,8 @@ class ServiceCrud:
 
         response = await client.table(SETTINGS.service_table).select("id").eq("id", service_id).execute()
 
-        return not(bool(response.data))
+
+        return bool(response.data)
     
     @classmethod
     async def get_input_services_by_service(cls, service_id: int) -> list[Product]:
@@ -94,3 +103,131 @@ class ServiceCrud:
             raise HTTPException(detail="No inputs found for this service", status_code=404)
 
         return [ProductCrud.get_product_by_id(service_input_id) for service_input_id in response.data[0].get("service_inputs", [])]
+    
+    @classmethod
+    async def create_service_input(cls, service_input_data: ServiceInputCreate) -> ServiceInput:
+        """Create a new service input."""
+        
+        client = await get_db_client()
+
+        response = await client.table(SETTINGS.service_inputs_table).insert(service_input_data.model_dump(mode="json")).execute()
+
+        if not(bool(response.data)):
+            raise HTTPException(detail="Failed to create service input", status_code=500)
+
+        return ServiceInput.model_validate(response.data[0])
+    
+    @classmethod
+    async def update_service_input(cls, service_id: int, product_id: int, fields: dict) -> ServiceInput:
+        """Update an existing service input."""
+        
+        client = await get_db_client()
+        
+        # check if service input exists
+        if not await cls.exist_service_input_by_service_id_and_product_id(service_id, product_id):
+            raise HTTPException(detail="Service input not found", status_code=404)
+
+        response = await client.table(SETTINGS.service_inputs_table).update(fields).eq("service_id", service_id).eq("product_id", product_id).execute()
+
+        if not(bool(response.data)):
+            raise HTTPException(detail="Failed to update service input", status_code=500)
+
+        return ServiceInput.model_validate(response.data[0])
+    
+    @classmethod
+    async def get_service_input_by_id(cls, service_input_id: int) -> ServiceInput:
+        """Retrieve a service input by ID."""
+        
+        client = await get_db_client()
+
+        response = await client.table(SETTINGS.service_inputs_table).select("*").eq("id", service_input_id).execute()
+
+        if not(bool(response.data)):
+            raise HTTPException(detail="Service input not found", status_code=404)
+
+        return ServiceInput.model_validate(response.data[0])
+    
+    @classmethod
+    async def update_service_input_by_id(cls, service_input_id: int, fields: dict) -> ServiceInput:
+        """Update a service input by ID."""
+        
+        client = await get_db_client()
+        
+        # check if service input exists
+        if not await cls.exist_service_input_by_id(service_input_id):
+            raise HTTPException(detail="Service input not found", status_code=404)
+
+        response = await client.table(SETTINGS.service_inputs_table).update(fields).eq("id", service_input_id).execute()
+
+        if not(bool(response.data)):
+            raise HTTPException(detail="Failed to update service input", status_code=500)
+
+        return ServiceInput.model_validate(response.data[0])
+    
+    @classmethod
+    async def get_service_input_by_service_id_and_product_id(cls, service_id: int, product_id : int) -> ServiceInput:
+        """Retrieve a service input by service ID and product ID."""
+        
+        client = await get_db_client()
+
+        response = await client.table(SETTINGS.service_inputs_table).select("*").eq("service_id", service_id).eq("product_id", product_id).execute()
+
+        if not(bool(response.data)):
+            raise HTTPException(detail="Service input not found", status_code=404)
+
+        return ServiceInput.model_validate(response.data[0])
+    
+    @classmethod
+    async def delete_service_input_by_service_id_and_product_id(cls, service_id: int, product_id: int) -> bool:
+        """Delete a service input by service ID and product ID."""
+        
+        client = await get_db_client()
+        
+        # check if service input exists
+        if not await cls.exist_service_input_by_service_id_and_product_id(service_id, product_id):
+            raise HTTPException(detail="Service input not found", status_code=404)
+
+        response = await client.table(SETTINGS.service_inputs_table).delete().eq("service_id", service_id).eq("product_id", product_id).execute()
+
+        if not(bool(response.data)):
+            raise HTTPException(detail="Failed to delete service input", status_code=500)
+
+        return True
+    
+    @classmethod
+    async def delete_service_input_by_id(cls, service_input_id: int) -> bool:
+        """Delete a service input by ID."""
+        
+        client = await get_db_client()
+        
+        # check if service input exists
+        if not await cls.exist_service_input_by_id(service_input_id):
+            raise HTTPException(detail="Service input not found", status_code=404)
+
+        response = await client.table(SETTINGS.service_inputs_table).delete().eq("id", service_input_id).execute()
+
+        if not(bool(response.data)):
+            raise HTTPException(detail="Failed to delete service input", status_code=500)
+
+        return True
+    
+    @classmethod
+    async def exist_service_input_by_service_id_and_product_id(cls, service_id: int, product_id: int) -> bool:
+        """Check if a service input exists by service ID and product ID."""
+        
+        client = await get_db_client()
+
+        response = await client.table(SETTINGS.service_inputs_table).select("id").eq("service_id", service_id).eq("product_id", product_id).execute()
+
+        return bool(response.data)
+    
+    @classmethod
+    async def exist_service_input_by_id(cls, service_input_id: int) -> bool:
+        """Check if a service input exists by ID."""
+        
+        client = await get_db_client()
+
+        response = await client.table(SETTINGS.service_inputs_table).select("id").eq("id", service_input_id).execute()
+
+
+        return bool(response.data)
